ProgressPost
TODO: Fix previous bugs, add ClientRPC, merge with BaseInventory
