+++ code/ui/Hud.cs
@@ -15,10 +15,10 @@ namespace TTTReborn.UI
             RootPanel.AddChild<ChatBox>();
             RootPanel.AddChild<VoiceList>();
             RootPanel.AddChild<GameTimer>();
-            RootPanel.AddChild<WeaponSelection>();
             RootPanel.AddChild<InfoFeed>();
             RootPanel.AddChild<InspectMenu>();
             RootPanel.AddChild<PlayerInfo>();
+            RootPanel.AddChild<WeaponSelection>();
         }
 
         [ClientRpc]
