+++ code/weapons/WeaponFunctions.cs
